Object hash is removed from custom graphics code.  From this version, it uses Long value as immutable ID.
git-svn-id: http://chianti.ucsd.edu/svn@20829 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ cytoscape/trunk/src/cytoscape/visual/customgraphic/URLImageCustomGraphicsParser.java
@@ -1,6 +1,7 @@
 package cytoscape.visual.customgraphic;
 
 import cytoscape.Cytoscape;
+import cytoscape.visual.customgraphic.impl.bitmap.URLImageCustomGraphics;
 
 /**
  * Create instance of URLImageCustomGraphics object from String.
@@ -35,7 +36,7 @@ public class URLImageCustomGraphicsParser implements CyCustomGraphicsParser {
 
 		final String imageName = entry[1];
 		CyCustomGraphics<?> cg = Cytoscape.getVisualMappingManager()
-				.getCustomGraphicsPool().getByID(Integer.parseInt(imageName));
+				.getCustomGraphicsPool().getByID(Long.parseLong(imageName));
 		cg.setDisplayName(entry[2]);
 		return cg;
 	}
